Fixup dynamodb portion of part1
The final state for part 1 section 2 shows the ``db.py``
that has the dynamodb table already added.  This is confusing
because it's not introduced until the next section.  There's no
actual code snippet that contains just this version so I inlined
the code instead of using the ``literalinclude``.

I also removed a duplicate step.  It was confusing because we just
copied the file a few steps back.